♻️ move Stat component to shared location
